requirements.txt
astropy
numpy
joblib
optuna
pandas
torch
plotly
matplotlib
torchvision
PyYAML
PyQt5
sbi
corner
zuko
SciencePlot